requirements.txt
# Core dependencies
pydantic>=2.0.0
typing-extensions>=4.0.0
python-dotenv>=1.0.0

# LLM provider clients
openai>=1.0.0
anthropic>=0.7.0
google-generativeai>=0.3.0
cohere>=4.0.0
requests>=2.28.0

# Optional dependencies for cost estimation
tiktoken>=0.5.0

# Development dependencies
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-cov>=4.0.0
black>=23.0.0
isort>=5.0.0
flake8>=6.0.0
mypy>=1.0.0 